Try to deploy on github

+++ ci/deploy.sh
@@ -1,5 +1,17 @@
 #!/usr/bin/env bash
 
+# We need to install dependencies only for Docker
+[[ ! -e /.dockerenv ]] && exit 0
+
+set -xe
+
+## Install global packages
+apt-get update -yqq
+
+
+
+
+
 # Install git.
 apt-get install git -yqq
